refactor: split transaction API module by responsibility
