+++ frontend/src/api/transactions/index.ts
@@ -1,0 +1,35 @@
+export type {
+  CreateTransactionPayload,
+  DailyCashFlow,
+  OutlierTransaction,
+  OverviewFilters,
+  TopCategorySpend,
+  Transaction,
+  TransactionFilters,
+  TransactionTag,
+  TransactionsOverview,
+  UpdateTransactionPayload,
+} from '@/api/transactions/types';
+
+export {
+  createTransaction,
+  deleteTransaction,
+  fetchTransaction,
+  fetchTransactionPage,
+  fetchTransactions,
+  fetchTransactionsOverview,
+  updateTransaction,
+} from '@/api/transactions/requests';
+
+export {
+  invalidateTransactionAccountData,
+} from '@/api/transactions/cache';
+
+export {
+  useCreateTransaction,
+  useDeleteTransaction,
+  useInfiniteTransactions,
+  useTransactions,
+  useTransactionsOverview,
+  useUpdateTransaction,
+} from '@/api/transactions/hooks';
